% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{add_html_page}
\alias{add_html_page}
\title{Create a page on Google Sites}
\source{
utils.R
}
\usage{
add_html_page(page_xhtml_source, page_parent_page_title = "",
  page_template_id = "", page_title = "", page_custom_url = "",
  overwrite = TRUE, site_domain = getOption("googlesites.site_domain"),
  site_name = getOption("googlesites.site_name"))
}
\arguments{
\item{page_xhtml_source}{A string representing the path to a local HTML file that can be successfully 
read into R as XML. Google SItes data must be valid XML so this check is performed by \code{xml2::read_xml()}}

\item{page_parent_page_title}{A string (optional) with the title of a webpage on the existing Google 
Site where this particular post will be created as a subpage}

\item{page_template_id}{A string (optional) with id of a template on the existing Google 
Site that this particular post will inherit}

\item{page_title}{A string (optional) that will be used as the page Title on the created post. If not
specified, then the title will be the same as the HTML file name, excluding .html and santitized to 
only include letters, numbers, hypens, or underscores.}

\item{page_custom_url}{A string (optional) that will be used as the page URL on the created post. If not
specified, then the url will be the same as the HTML file name, excluding .html and santitized to 
only include letters, numbers, hypens, or underscores.}

\item{overwrite}{A logical indicating whether to overwrite any existing pages on the Google Site with the
same title or url. Matches on url are prioritized above matches with same title, meaning that if a page with 
the same title and another page with the same url are found, then we will overwrite the one with the same url
since, technically, Google Sites pages with the same title are allowed, even though we would prefer to keep those
unique under each parent page}

\item{site_domain}{A string representing 'site' or the domain of your Google Apps hosted domain (e.g. example.com)}

\item{site_name}{A string representing the webspace name of your site; found in the Site's URL (e.g. myCoolSite)}
}
\description{
This function takes metadata and a file path to an HTML
document and pushes that HTML up to a Google Site
}
\examples{
\dontrun{
# add a page
test_html <- system.file("inst", "test.html", package="googlesites")
add_html_page(page_xhtml_source = test_html)

# add a page and overwrite the existing one in event it already exists
sample_html <- system.file("extdata", "sample-doc.html", package="googlesites")
add_html_page(page_xhtml_source = sample_html,
              page_title = 'Post with GIF',
              page_custom_url = 'post-with-gif', 
              overwrite=TRUE)
}
}
\keyword{api}
\keyword{create}
\keyword{documentation}
\keyword{googlesites}
\keyword{page}
\keyword{webpage}
\keyword{wiki}
