% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{delete_content}
\alias{delete_content}
\title{Delete a page on Google Sites}
\usage{
delete_content(id, verbose = TRUE)
}
\arguments{
\item{id}{A string representing the URL that identifies the entry id of content 
to delete. This should be an absolute URL looking like:
https://sites.google.com/feeds/content/mydomain/mysite/6477233125232797597. You can find this URL 
by using the \code{\link{find_content}} function.}

\item{verbose}{A logical indicating whether to print messages}
}
\description{
This function takes a title, url, or Id of a page, then deletes it from 
a Google Sites
}
\examples{
\dontrun{
# delete a webpage with title My Report
delete_content(id = find_content(value_to_match='My Report',
                                 field_to_match='title')$id)
            
# delete an attachment 
delete_content(id = find_content(value_to_match='rmarkdown-cheatsheet-2.0',
                                 field_to_match='title', 
                                 content_category='attachment')$id)
}
}
\keyword{announcement}
\keyword{api}
\keyword{attachment}
\keyword{delete}
\keyword{documentation}
\keyword{googlesites}
\keyword{webpage}
\keyword{wiki}
