% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exponential_backoff_retry}
\alias{exponential_backoff_retry}
\title{Retry a Function With Exponential Backoff}
\usage{
exponential_backoff_retry(expr, n = 5, verbose = FALSE)
}
\arguments{
\item{expr}{an expression to be evaluated with exponential backoff}

\item{n}{an integer or number indicating the number of times to execute the function
before completely failing}

\item{verbose}{a \code{logical} indicating whether to print messages when invoking
the retry attempts}
}
\value{
the result of the evaluated expression
}
\description{
This function is an internal helper that will retry a function 
will exponential wait times in case there are issues with the API
}
\examples{
\dontrun{
# use with an API call that you can't afford to error because of intermittent uptime
myFun <- function(){
 x <- runif(1)
 if(x < .6){
   stop('less than 3/5')
 }
 return(x)
}
exponential_backoff_retry(expr = myFun(), verbose = TRUE, n = 5)
}
}
\keyword{internal}
