% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sites_DELETE}
\alias{sites_DELETE}
\title{sites_DELETE}
\usage{
sites_DELETE(url, gdata_version = getOption("googlesites.gdata_version"),
  host = getOption("googlesites.host"))
}
\description{
sites_DELETE
}
\keyword{internal}
