% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sites_POST}
\alias{sites_POST}
\title{sites_POST}
\usage{
sites_POST(url, body, content_type,
  gdata_version = getOption("googlesites.gdata_version"),
  host = getOption("googlesites.host"))
}
\description{
sites_POST
}
\keyword{internal}
