% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sites_PUT}
\alias{sites_PUT}
\title{sites_PUT}
\usage{
sites_PUT(url, body, content_type,
  gdata_version = getOption("googlesites.gdata_version"),
  host = getOption("googlesites.host"))
}
\description{
sites_PUT
}
\keyword{internal}
