% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{delete_content}
\alias{delete_content}
\title{Delete a page on Google Sites}
\usage{
delete_content(id)
}
\arguments{
\item{value_to_match}{A string that should match an entry on the target google site}

\item{field_to_match}{A string representing the title that the attachment will be created with}

\item{parent_page_id}{A string representing the url that identifies the entry id of a webpage
where the uploaded file will be attached to typically formatted as /feeds/content/domainName/siteName/PARENT_ENTRY_ID}

\item{content_category}{A string or a collection of strings that specify which types of entries to search against. If 
none are provided then all entries are scanned. This argument acts as a filter to speed up the search process. Accepted 
values are announcement, announcementspage, attachment, comment, filecabinet, listitem, listpage, webpage, webattachment, template}

\item{site_domain}{A string representing 'site' or the domain of your Google Apps hosted domain (e.g. example.com)}

\item{site_name}{A string representing the webspace name of your site; found in the Site's URL (e.g. myCoolSite)}
}
\description{
This function takes a title, url, or Id of a page, then deletes it from 
a Google Sites
}
\examples{
\dontrun{
# delete a webpage with title My Report
delete_content(id = find_content(value_to_match='My Report',
                                 field_to_match='title')$id)
            
# delete an attachment 
delete_content(id = find_content(value_to_match='rmarkdown-cheatsheet-2.0',
                                 field_to_match='title', 
                                 content_category='attachment')$id)
}
}
\keyword{announcement}
\keyword{api}
\keyword{attachment}
\keyword{delete}
\keyword{documentation}
\keyword{googlesites}
\keyword{webpage}
\keyword{wiki}
